snd_requirements.txt
humanize==4.5.0
python-dotenv==0.20.0